pyproject.toml
[project]
name = "multi-prompt-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit>=1.0.11",
    "livekit-agents[cartesia,deepgram,elevenlabs,openai,silero,turn-detector]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
]
